Squashed some compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@3978 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ lib/init_rs.c
@@ -38,7 +38,7 @@ void *INIT_RS(int symsize,int gfpoly,int fcr,int prim,
   int i, j, sr,root,iprim;
 
   /* Check parameter ranges */
-  if(symsize < 0 || symsize > 8*sizeof(DTYPE))
+  if(symsize < 0 || symsize > (int)(8*sizeof(DTYPE)))
     return NULL; /* Need version with ints rather than chars */
 
   if(fcr < 0 || fcr >= (1<<symsize))
